feat: BI-5792 Upgrade MySQL version to 8.0.40 (#729)
* Upgrade MySQL version to 8.0.40

* Update MySQL connector to use DOUBLE type for dates (because of FLOAT overflow)

* Update MySQL dialect mentions to version 8.0.40 across tests and definitions

* Update MySQL version references to 8.0.40 in dialects and localization files

* Compiled licalization

* Fix backward compatibility issues

* Revert "Fix backward compatibility issues"

This reverts commit 47fcf8339ff0674d7daf131a0d400d397b2953f8.

* Revert "Compiled licalization"

This reverts commit bddda72706c3e7d3f2651e045b2e576ef6273c66.

* Revert "Update MySQL version references to 8.0.40 in dialects and localization files"

This reverts commit e266bd05ae85c1e6d67723a5fea3b7769a4894a6.

* Revert "Update MySQL dialect mentions to version 8.0.40 across tests and definitions"

This reverts commit 93c7fce734f06558f39907c5209b1cc60096f53f.

---------

+++ lib/dl_connector_mysql/dl_connector_mysql/formula/definitions/functions_type.py
@@ -60,17 +60,17 @@ DEFINITIONS_TYPE = [
     ),
     base.FuncFloatFromDate(
         variants=[
-            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.FLOAT)),
+            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.DOUBLE)),
         ]
     ),
     base.FuncFloatFromDatetime(
         variants=[
-            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.FLOAT)),
+            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.DOUBLE)),
         ]
     ),
     base.FuncFloatFromGenericDatetime(
         variants=[
-            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.FLOAT)),
+            V(D.MYSQL, lambda value: sa.cast(sa.func.UNIX_TIMESTAMP(value), sa_mysql.DOUBLE)),
         ]
     ),
     # genericdatetime
